Change the page's name
The page's name was by default the page file's name before ,
which was Spaghetti.html. and  i decided to change it ,
not just because it wasn't meaningful or informative at all ,
it is just because i thought i'd change it to a name that is more
detailed and look nicer and cleaner with the removal of
the extension part , so it would look more fitting and simpler for
users.